django/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
commonmark==0.9.1
Django==3.2.9
django-mysql==4.2.0
django-reverse-admin==2.9.6
djangorestframework==3.12.4
idna==3.3
mysqlclient==2.0.3
pip-search==0.0.7
Pygments==2.10.0
pytz==2021.3
requests==2.26.0
rich==10.14.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7
uWSGI==2.0.20
